pyproject.toml
[build-system]
requires = ["scikit-build-core", "pybind11"]
build-backend = "scikit_build_core.build"

[project]
name = "helicopter"
version = "0.1.0"
dependencies = [
    "numpy",
    "numpy-quaternion",
    "pandas",
    "filterpy",
    "matplotlib",
    "scipy"]

[tool.scikit-build]
wheel.install-dir = "helicopter"
cmake.targets = ["helicopter_cpp"]
editable.mode = "inplace"